Separate Group object
